+++ multi_final/front/workspace_react/src/page/communityGroup/GroupBoardDetail.js
@@ -26,10 +26,13 @@ const GroupBoardDetail = () => {
   
   }
   function Delete(no){
-       
-    axios.delete(`http://localhost:8085/deleteGroup/${no}`)
+    if(window.confirm("게시글을 삭제할까요?")){
+      axios.delete(`http://localhost:8085/deleteGroup/${no}`)
          .then(navigate('/communityGroup')).catch(err=>console.log(err))
-      }
+    } else {
+      console.log("삭제 취소");
+    }
+  }
